# openapi.yaml
openapi: 3.1.0
info:
  title: Self Agent ID API
  version: 1.0.0
  description: >
    On-chain AI agent identity registry with proof-of-human verification.
    Register, deregister, and query agents — no API keys required.
    Sessions use encrypted tokens with a 30-minute TTL.
  contact:
    name: Self Protocol
    url: https://github.com/selfxyz/self-agent-id
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://self-agent-id.vercel.app
    description: Production
...
security: []
...
components:
  securitySchemes:
    selfAgentAuth:
      type: apiKey
      in: header
      name: x-self-agent-signature
      description: >
        ECDSA signature produced by `SelfAgent.signRequest()`.
        Must be accompanied by `x-self-agent-address` and
        `x-self-agent-timestamp` headers.
    aaProxyToken:
      type: apiKey
      in: header
      name: x-aa-proxy-token
      description: >
        Short-lived token issued by POST /api/aa/token.
        Required for bundler and paymaster proxy endpoints.
  parameters:
    chainIdPath:
      name: chainId
      in: path
      required: true
      schema:
        type: integer
        enum: [42220, 11142220]
      description: "42220 (Celo Mainnet) or 11142220 (Celo Sepolia Testnet)"
    agentIdPath:
      name: agentId
      in: path
      required: true
      schema:
        type: integer
      description: On-chain agent token ID
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Credentials:
      type: object
      properties:
        nationality:
          type: string
          description: ISO 3166-1 alpha-3 country code
        olderThan:
          type: integer
        ofac:
          type: array
          items:
            type: boolean
          description: "[isOnSdnList, isOnConsolidatedList, isOnOfacList]"
    AgentInfoResponse:
      type: object
      properties:
        agentId:
          type: integer
        chainId:
          type: integer
        agentAddress:
          type: string
        isVerified:
          type: boolean
        proofProvider:
          type: string
        verificationStrength:
          type: integer
        strengthLabel:
          type: string
        credentials:
          $ref: "#/components/schemas/Credentials"
        registeredAt:
          type: integer
          description: Unix timestamp
        network:
          type: string
...
    AgentToAgentResponse:
      type: object
      properties:
        verified:
          type: boolean
        demoAgent:
          type: object
          properties:
            address:
              type: string
            agentId:
              type: string
            verified:
              type: boolean
        callerAgent:
          type: object
          properties:
            address:
              type: string
            agentId:
              type: string
            verified:
              type: boolean
        sameHuman:
          type: boolean
        verificationCount:
          type: integer
        uniqueAgents:
          type: integer
        message:
          type: string
...
paths:
...
  /api/agent/info/{chainId}/{agentId}:
    get:
      tags: [Query]
      summary: Get agent details
      operationId: getAgentInfo
      description: >
        Returns full agent information: address, verification status,
        proof provider, credentials, and registration timestamp.
      parameters:
        - $ref: "#/components/parameters/chainIdPath"
        - $ref: "#/components/parameters/agentIdPath"
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInfoResponse"
              example:
                agentId: 5
                chainId: 11142220
                agentAddress: "0x83fa4380903fecb801F4e123835664973001ff00"
                isVerified: true
                proofProvider: "0x8e248DEB0F18B0A4b1c608F2d80dBCeB1B868F81"
                verificationStrength: 2
                strengthLabel: Standard
                credentials:
                  nationality: GBR
                  olderThan: 18
                  ofac: [false, false, false]
                registeredAt: 1740000000
                network: testnet
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/demo/agent-to-agent:
    post:
      tags: [Demo]
      summary: Agent-to-agent verification
      operationId: demoAgentToAgent
      description: >
        Demo agent endpoint that verifies the caller, checks
        `sameHuman()` on-chain, and signs its response so the caller
        can verify it came from the demo agent.
      security:
        - selfAgentAuth: []
      parameters:
        - name: network
          in: query
          schema:
            type: string
            enum: [celo, celo-sepolia]
      responses:
        "200":
          description: Mutual verification result (response headers include demo agent signature)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentToAgentResponse"
        "401":
          description: Missing agent authentication headers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Agent verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Replay detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...